refactor: "tools" event handling (#1308)
* refactor: tools event handling

* chore: separate tools to individual modules
